requirements.txt
Django==2.0.7
djangorestframework==3.8.1
djangorestframework-camel-case==0.2.0
psycopg2==2.7.5
pytz==2018.5
